backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
google-generativeai>=0.7.2
#openai-whisper==20231117
langdetect==1.0.9
google-cloud-texttospeech==2.15.1
celery==5.3.4
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0